Alteração de algumas funções

+++ backend/src/main/java/kn/kommute/app/service/RideService.java
@@ -5,11 +5,16 @@ import kn.kommute.app.mapper.RideMapper;
 import kn.kommute.app.model.Ride;
 import kn.kommute.app.model.User;
 import kn.kommute.app.repository.RideRepository;
+import kn.kommute.app.security.CustomUserDetails;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+
 
 import java.util.*;
 
@@ -35,7 +40,7 @@ public class RideService {
         List<Ride> rides = rideRepository.findAll();
         List<RideDTO> response = new ArrayList<>();
         for (Ride ride : rides) {
-            response.add(rideMapper.toSummaryDTO(ride));
+            response.add(rideMapper.toRideDTO(ride));
         }
         return response;
     }
@@ -51,17 +56,17 @@ public class RideService {
         ride.setTotalCarpoolers(ride.getTotalCarpoolers() + 1);
         Ride savedRide = rideRepository.save(ride);
 
-        return rideMapper.toParticipationDTO(savedRide);
+        return rideMapper.toRideDTO(savedRide);
     }
 
     public Optional<RideDTO> findRideById(Long id) {
         return rideRepository.findById(id)
-                .map(ride -> rideMapper.toSummaryDTO(ride));
+                .map(ride -> rideMapper.toRideDTO(ride));
     }
 
     public Optional<RideDTO> findRideByOrigin(String origin) {
         return rideRepository.findByOrigin(origin)
-                .map(ride -> rideMapper.toSummaryDTO(ride));
+                .map(ride -> rideMapper.toRideDTO(ride));
     }
 
     public RideDTO updateRide(Long id, Ride updatedRide, User userRequest) {
@@ -79,12 +84,31 @@ public class RideService {
         ride.setMaxUsers(updatedRide.getMaxUsers());
 
         Ride savedRide = rideRepository.save(ride);
-        return rideMapper.toSummaryDTO(savedRide);
+        return rideMapper.toRideDTO(savedRide);
     }
 
+    public Long getAuthenticatedUserId() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new RuntimeException("User not authenticated");
+        }
 
-    public void deleteRide(Long id) {
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        return userDetails.getId();
+    }
+
+    public void deleteById(Long id) {
+
+        Long currentUserId = getAuthenticatedUserId();
+
+        Ride ride = rideRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Ride not found"));
+
+        if (!ride.getOwner().getId().equals(currentUserId)) {
+            throw new RuntimeException("You do not have permission to delete this ride");
+        }
         rideRepository.deleteById(id);
+
     }
 }
